# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a sample Private API",
    "version": "3.0.0",
    "title": "Node.js template for Private APIs",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "contact@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],  
  "paths": {    
...
    "/v1/attributes": {
      "parameters": [                
        {
          "$ref": "#/components/parameters/requestIdParam"
        }
      ],
      "get": {
        "summary": "Filter attributes",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/projectIdParam"
          },
          {
            "$ref": "#/components/parameters/tagsParam"
          },
          {
            "$ref": "#/components/parameters/deletedsParam"
          }
        ],
        "tags": ["attributes"]
      },
      "post": {
        "summary": "Create a new attribute",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAttribute",
                "required": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["attributes"]
      }
    },
    "/v1/attributes/{attributeId}": {
      "parameters": [
        {
          "in": "path",
          "name": "attributeId",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "required": true
        },
        {
          "$ref": "#/components/parameters/requestIdParam"
        }
      ],
      "get": {
        "summary": "Retrieve an existing attribute",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/components/parameters/deletedsParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["attributes"]
      },
      "put": {
        "summary": "Update an existing attribute",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAttribute",
                "required": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["attributes"]
      },
      "delete": {
        "summary": "Delete an existing attribute",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["attributes"]
      }
    },
    "/v1/services": {
      "get": {
        "summary": "Filter services",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/projectIdParam"
          },
          {
            "$ref": "#/components/parameters/searchParam"
          },
          {
            "$ref": "#/components/parameters/tagsParam"
          },
          {
            "$ref": "#/components/parameters/deletedsParam"
          },
          {
            "$ref": "#/components/parameters/requestIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Offset": {
                "type": "integer",
                "description": "Skipped records"
              },
              "X-Limit": {
                "type": "integer",
                "description": "Max returned records per request"
              },
              "X-Total-Count": {
                "type": "integer",
                "description": "Total found records"
              }
            },
            "schema": {
              "$ref": "#/components/ApiResponse"
            },
            "example": {
              "error": null,
              "data": {
                "count": 284,
                "edges": [
                  {
                    "id": "4578sad-...",
                    "name": "Service 1",
                    "createdAt": "2021-04-29T12:28:00-03:00"
                  },
                  {
                    "id": "4578sad-...",
                    "name": "Service 2",
                    "createdAt": "2021-04-29T12:28:00-03:00"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["services"]
      },
      "post": {
        "summary": "Create a new service",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": true,
                "$ref": "#/components/Service"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["services"]
      }
    },
    "/v1/services/{serviceId}": {
      "parameters": [
        {
          "name": "serviceId",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "required": true
        },
        {
          "$ref": "#/components/parameters/deletedsParam"
        }
      ],
      "get": {
        "summary": "Retrieve an existing service",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/components/parameters/deletedsParam"
          }      
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["services"]
      },
      "put": {
        "summary": "Update an existing service",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": true,
                "$ref": "#/components/CreateServiceBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["services"]
      },
      "delete": {
        "summary": "Delete an existing service",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["services"]
      }
    },
    "/v1/service-attributes": {
      "put": {
        "summary": "Update service attribute value",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": true,
                "$ref": "#/components/SetServiceAttributeValueBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["values"]
      }
    },
    "/v1/service-attributes/bulk": {
      "put": {
        "summary": "Bulk update service attribute value",
        "consumes": ["application/json", "application/x-ndjson"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": true,
                "properties": {
                  "items": {
                    "type": "array",
                    "required": true,
                    "items": {
                      "$ref": "#/components/SetServiceAttributeValueBody"
                    }
                  }
                }
              }
            },
            "application/x-ndjson": {
              "description": "New line delimited json",
              "schema": {
                "type": "string",
                "required": true,
                "example": "{\"serviceId\": \"ac8d5ac1-aa03-4850-8e2d-3916c2a8d5c3\", \"attributeId\": \"2aaeea02-d878-442e-87c5-ac2e6afe0294\", \"attributeValue\": \"Attribute 1 value\"}\r\n{\"serviceId\": \"ac8d5ac1-aa03-4850-8e2d-3916c2a8d5c3\", \"attributeId\": \"ceaeeb4d-1949-405b-b1e1-55f29f9138d1\", \"attributeValue\": \"Attribute 2 value\"}\r\n{\"serviceId\": \"ff7355e4-3829-4b1e-9e13-5d664f0eba7b\", \"attributeId\": \"2aaeea02-d878-442e-87c5-ac2e6afe0294\", \"attributeValue\": \"Attribute 1 value\"}\r\n{\"serviceId\": \"ff7355e4-3829-4b1e-9e13-5d664f0eba7b\", \"attributeId\": \"ceaeeb4d-1949-405b-b1e1-55f29f9138d1\", \"attributeValue\": \"Attribute 2 value\"}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["values"]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "components": {
    "parameters": {
      "requestIdParam": {
        "in": "header",
        "name": "X-Request-Id",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "required": false,
        "description": "ID of request"
      },
      "projectIdParam": {
        "in": "query",
        "name": "projectId",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "ID of project"
      },
      "offsetParam": {
        "in": "query",
        "name": "from",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "nullable": true
        },        
        "required": false,
        "description": "Records count to skip **(default=0)**"
      },
      "limitParam": {
        "in": "query",
        "name": "size",        
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": null
        },
        "required": false,
        "description": "Records per request **(default=20)**"
      },
...
      "searchParam": {
        "in": "query",
        "name": "q",
        "type": "string",
        "required": false,
        "description": "Search by record name"
      },
...
      "tagsParam": {
        "in": "query",
        "name": "tags",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "keypair",
            "example": "name=value"
          }
        },
        "style": "simple",
        "explode": false,
        "required": false,
        "description": "URI encoded tags (example: \"name1=value1,name2=value2\")"
      },
      "deletedsParam": {
        "in": "query",
        "name": "deleteds",
        "schema": {
          "type": "boolean"
        },
        "required": false,
        "description": "Enable/display deleted records"
      }
    },
    "schemas": {
...
      "PutAttribute": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid",
            "required": true
          },
          "name": {
            "type": "string",
            "required": true,
            "example": "My attribute name"
          },
          "format": {
            "type": "string",
            "enum": ["text", "integer", "float", "date", "date-time"],
            "default": "text"
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          "isSearchable": {
            "type": "boolean",
            "default": false
          },
          "isDisabled": {
            "type": "boolean",
            "default": false
          },
          "isPinned": {
            "type": "boolean",
            "default": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "default": null
          }
        }
      }
    },
...
  },
...
}